+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 - Initial Open Source release.
 ## 1.0.1
 - update the overlapping and installation improvements
+## 1.0.2
+- Added features
+- multiselect functionality
+- reactive UI on selection
+- Display multiple selected items inside the dropdown field.